refactoring search box as a component
Refactoring search box as a component
