+++ src/components/Navbar/Navbar.component.jsx
@@ -3,8 +3,8 @@ import './Navbar.styles.css';
 import { Link, useHistory } from 'react-router-dom';
 import styled from 'styled-components';
 import { useAuth } from '../../providers/Auth';
-import { VideoContext } from '../../providers/Video';
 import { ThemeContext } from '../../providers/Theme';
+import SearchBox from '../SearchBox/SearchBox.component';
 
 const NavBarHeader = styled.header`
   display: flex;
@@ -30,65 +30,6 @@ const NavBarItem = styled.div`
   cursor: pointer;
   vertical-align: middle;
   display: flex;
-`;
-
-const NavBarSearchBox = styled.div`
-  margin-right: auto;
-`;
-
-const Cover = styled.div`
-  top: 50%;
-  left: 0;
-  right: 0;
-  width: 400px;
-  padding: 7px;
-  background-color: ${(props) => props.theme.variant};
-  border-radius: 8px;
-  box-shadow: 0 4px 22px ${(props) => props.theme.variant}, 0 0 0 10px #ffffffeb;
-  transform: scale(0.6);
-`;
-
-const TB = styled.div`
-  display: table;
-  width: 100%;
-`;
-
-const TD = styled.div`
-  display: table-cell;
-  vertical-align: middle;
-`;
-
-const SearchInput = styled.input`
-width: 100%;
-height: 50px;
-font-size: 30px;
-line-height: 1;
-border: none;
-color: #fff;
-font-family: Nunito;
-padding: 0;
-margin: 0;
-border: 0;
-background-color: transparent;
-&:focus {
-  outline: none;
-}
-&:placeholder {
-  color: #1c5476;
-`;
-
-const SearchButton = styled.button`
-  color: #fff;
-  font-family: Nunito;
-  padding: 0;
-  margin: 0;
-  border: 0;
-  background-color: transparent;
-  position: relative;
-  display: block;
-  width: 84px;
-  height: 55px;
-  cursor: pointer;
 `;
 
 const CheckBoxLabel = styled.label`
@@ -144,7 +85,6 @@ const DarkSide = styled.label`
 
 function NavBar() {
   const history = useHistory();
-  const { searchTerm, setSearchTerm, fetchVideos } = useContext(VideoContext);
   const { isDarkTheme, setIsDarkTheme } = useContext(ThemeContext);
 
   const { authenticated, logout } = useAuth();
@@ -155,49 +95,14 @@ function NavBar() {
     history.push('/');
   }
 
-  const handleChangeSearchInput = (value) => {
-    setSearchTerm(value);
-  };
-
-  const handleClickSearch = () => {
-    fetchVideos();
-  };
-
   const handleChangeTheme = () => {
     setIsDarkTheme(!isDarkTheme);
-  };
-
-  const handleKeyPress = (event) => {
-    if (event.key === 'Enter' || event.keyCode === 13) {
-      fetchVideos();
-    }
   };
 
   return (
     <NavBarHeader>
       <NavBarTitle>Academy project</NavBarTitle>
-      <NavBarSearchBox>
-        <Cover id="cover">
-          <TB>
-            <TD>
-              <SearchInput
-                type="text"
-                placeholder="Search"
-                required
-                value={searchTerm}
-                onChange={(event) => handleChangeSearchInput(event.target.value)}
-                onKeyPress={(event) => handleKeyPress(event)}
-              />
-            </TD>
-            <TD id="s-cover">
-              <SearchButton type="submit" onClick={handleClickSearch}>
-                <div id="s-circle" />
-                <span />
-              </SearchButton>
-            </TD>
-          </TB>
-        </Cover>
-      </NavBarSearchBox>
+      <SearchBox></SearchBox>
       <NavBarItem>
         <Link to="/">Home</Link>
       </NavBarItem>
