+++ src/components/SearchBox/index.js
@@ -1,0 +1,1 @@
+export { default } from './SearchBox.component';
